Move hard files
