+++ nni/nas/execution/trial_entry.py
